Adjusted attribute trait to be a generic properties trait
Relationship and attribute methods now exist on one trait
